few changes i dont remember exactly

+++ libdllcore/plugin/DllFolderPlugin.cpp
@@ -131,7 +131,7 @@ SYSTEMTIME systemTimeAdd(SYSTEMTIME s, double seconds) {
 	memcpy(&u, &f, sizeof(u));
 
 	const double c_dSecondsPer100nsInterval = 100. * 1.E-9;
-	u.QuadPart += seconds / c_dSecondsPer100nsInterval;
+	u.QuadPart += ULONGLONG(seconds / c_dSecondsPer100nsInterval);
 
 	memcpy(&f, &u, sizeof(f));
